Revert "Se cambio el modo de cargar la fuente en memoria, eliminando el cast inecesario a byte."
This reverts commit 1573ffbc60676dc9225fc7c057280f915e0bb9b7.

+++ src/main/java/mx/uach/fing/chip8/Memory.java
@@ -91,7 +91,7 @@ public class Memory {
 
         // Copia la fuente en memoria.
         for (int i = 0; i < fonts.length; i++) {
-            this.ram[FONT_INIT + i] = fonts[i] & 0xFF;
+            this.ram[FONT_INIT + i] = (byte) (fonts[i] & 0xFF);
         }
     }
